proxy/requirements.txt
watchdog==3.0.0
http-parser==0.9.0
pymongo==4.6.3
dpkt==1.9.8
python-dotenv==1.0.1
